% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{Arrange_Table}
\alias{Arrange_Table}
\title{Arrange Table Function - Sort and filter columns of a data frame or matrix by a specified vector}
\usage{
Arrange_Table(Mat, FilterList, mode = "V2")
}
\arguments{
\item{Mat}{Target matrix or df.}

\item{FilterList}{Accord which to sort and filter columns}

\item{mode}{in Version 5.6 update, new "V2" mode is enabled. you can choose the V1 mode to change to the origin mode.}
}
\value{
a Matrix
}
\description{
Arrange Table Function - Sort and filter columns of a data frame or matrix by a specified vector
}
\author{
Zhiming Ye
}
