% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{lsEnv}
\alias{lsEnv}
\title{View Environment object and return a list}
\usage{
lsEnv(envir = parent.frame())
}
\arguments{
\item{envir}{Show which environment, default is `parent.frame()`}
}
\value{
a list
}
\description{
View Environment object and return a list
}
\author{
Zhiming Ye
}
