% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{PreviewDF}
\alias{PreviewDF}
\title{Previewing data table elements like rownames or colnames using Ace Editor}
\usage{
PreviewDF(x, n_lines = 7L)
}
\arguments{
\item{x}{A data.frame, tibble ...}

\item{n_lines}{display n rows}
}
\value{
A shiny page
}
\description{
Display `rownames()`, `colnames()` and `head()` result in Ace Editor. Ace Editor provides searching function and basic highlight functions. You can better view it in large tables
}
\author{
Zhiming Ye
}
