% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{Print.Char}
\alias{Print.Char}
\title{An enhanced method to print character vector to the console}
\usage{
Print.Char(CharacterCollection, Type = "c", return = F)
}
\arguments{
\item{CharacterCollection}{A character vector which will be print.}

\item{Type}{A character. Can be "c", "tab" or "plus". c means comma, tab means print in multi-lines, and plus means to separate with "+"}

\item{return}{if return=F, only print in console. Set TRUE to return an object.}
}
\description{
An enhanced method to print character vector to the console
}
\author{
Zhiming Ye
}
