% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{ViewDF}
\alias{ViewDF}
\title{Viewing tables in browser}
\usage{
ViewDF(x, n = 10000)
}
\arguments{
\item{x}{Data frame}

\item{n}{Display n lines}
}
\value{
A shiny website
}
\description{
Viewing tables in browser
}
\author{
Zhiming Ye
}
