% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{ViewEnv}
\alias{ViewEnv}
\title{View Environment object in shiny}
\usage{
ViewEnv(envir = parent.frame())
}
\arguments{
\item{envir}{Show which environment, default is `parent.frame()`}
}
\value{
a shiny website
}
\description{
View Environment object in shiny
}
\author{
Zhiming Ye
}
