% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{ViewObj}
\alias{ViewObj}
\title{Viewing Attributes of R Objects Printed in the Terminal in browser}
\usage{
ViewObj(
  x,
  selfExpr = F,
  pattern1 = NULL,
  pattern2 = NULL,
  pattern3 = NULL,
  move = 1,
  removeBlank = T,
  warning = T
)
}
\arguments{
\item{x}{Object, an S4 Object is tested. If provide other type, please consider that you should modify patterns to filtering and visualize. Or you can pass an self-defined expression to x, when set `selfExpr=F`}

\item{selfExpr}{In default, `capture.output(str(x,strict.width="cut"))` is used for generate the attribute of object. If disable, you can pass your self expression in x}

\item{pattern1}{Pattern `"^(( ..)+)"` is used to filter ` .. .. ..` like pattern and constructing tree-like object, pass to `gregexpr()` function.}

\item{pattern2}{Pattern to use to fetch object name}

\item{pattern3}{Pattern to used to fetch other parts}

\item{move}{Numeric, to adjust where the tree is start. You can set to 1 to avoid the root.}

\item{removeBlank}{In S4 object, the printed result starts with blank, set to TRUE to remove it}
}
\value{
A shiny website
}
\description{
This function is designed to parse the attributes of objects printed by the `str()`  function. In RStudio, the Environment pane provides a good overview of objects. However, VSCode does not offer similar support. For some complex S4 objects, providing an appropriate preview method is essential.
}
\author{
Zhiming Ye
}
